# Allow squashing f-contiguous axes for faster reduction