`_reduce_dims()` (and `ndarray.reduced_view()`) can only squash C-contiguous dimensions.
This PR sorts the axes so that they can be squashed even if they're F-contiguous.

Benchmark code using [cupy-perf](https://github.com/niboshi/cupy-perf/):
```py
import cupy
import cupy_perf


class Perf1(cupy_perf.PerfCases):

    def setUp(self):
        shape = (100,) * 3
        self.a = cupy.empty(shape, 'float32')
        self.b = cupy.empty(shape, 'float32').T

    @cupy_perf.attr(n=100)
    def perf_ccont(self):
        self.a.sum(axis=(1, 2))

    @cupy_perf.attr(n=100)
    def perf_fcont(self):
        self.b.sum(axis=(0, 1))


cupy_perf.run(__name__)
```

Result:
```
# Before                CPU Time                                    GPU Time
ccont               :    22.412 us   +/- 2.332 (min:   20.623) us    896.698 us   +/- 5.373 (min:  875.040) us
fcont               :    26.003 us   +/-17.168 (min:   17.974) us   2862.353 us   +/-16.635 (min: 2845.184) us

# This PR
ccont               :    26.467 us   +/- 1.182 (min:   25.313) us    901.093 us   +/- 4.774 (min:  882.048) us
fcont               :    26.440 us   +/- 1.400 (min:   24.819) us    899.893 us   +/- 4.466 (min:  883.712) us
```

~TODO: I think output arrays should be transposed.~
Currently targeting only 0- or 1-dim output.