@niboshi 2019-12-18
Redone benchmark:
```
# Before                CPU Time                                    GPU Time
ccont               :    28.263 us   +/-14.547 (min:   19.935) us    904.360 us   +/-12.922 (min:  876.288) us
fcont               :    24.622 us   +/- 7.184 (min:   20.392) us   1829.935 us   +/- 8.798 (min: 1813.760) us

# This PR
ccont               :    27.192 us   +/- 1.770 (min:   24.683) us    900.502 us   +/- 4.682 (min:  872.672) us
fcont               :    25.762 us   +/- 2.941 (min:   23.022) us    491.138 us   +/- 3.892 (min:  471.424) us
```