@niboshi 2019-12-18
`block_size`, `block_stride`, `out_block_num`

`ccont`: 512 1 100
`fcont`:  512 32 4

This explains why `fcont` is faster.