@niboshi 2019-12-18
>This explains why `fcont` is faster than `ccont`.

Or maybe because `fcont` squashes input arrays more than `ccont`:

`in_shape`, `out_shape` after `_reduce_dims()`

`ccont`: (10000, 100) (100,)
`fcont`: (1000000,) (100,)