@niboshi 2019-12-18
Axes in `in_shape` of `ccont` case is separated as (10000, 100) because the implementation always moves the reduction axes (10000) to the left, and output axes (100) to the right.

This transposition makes the input array non-contiguous.

In `fcont` case, this transposition keeps the f-contiguity of the input array, thus squashable.