@niboshi 2019-12-18
(Some of the) Tests are not unrelated, but  OK to separate a PR.