@niboshi 2019-12-18
Rebased. PTAL