@asi1024 2019-12-19
Confirmed performance improvements: [script](https://gist.github.com/asi1024/e8e2b77533fbe2451205f3de70b2fdba) [results](https://docs.google.com/spreadsheets/d/1Mz8psOUJTOjIpr3X_KEX6cJYoayOd3DRypZUv3h4gOM/edit#gid=0)